gt_inference::spec

Trait Mergeable

pub trait Mergeable: SpecCore {
    // Required methods
    fn merge_groups(&mut self, s: Group, t: Group, sc: &mut Self::Scratch);
    fn virtual_merge(
        &self,
        s: Group,
        t: Group,
        ea: &Self::Args,
        sc: &mut Self::Scratch,
    ) -> f64;
}

Group merging. Capability as trait presence.

Replaces the duplicated pair has_merge on the spec (spec_base.hh:50) and _has_merge on Imp (:115) and again on the loop states (merge_split.hh:99, multilevel.hh:114) – two independent booleans for one capability, at two scopes, which a derived spec can set inconsistently. if constexpr (_has_merge) becomes a where-clause, and there is no second flag to disagree with the first.

Required Methods

fn merge_groups(&mut self, s: Group, t: Group, sc: &mut Self::Scratch)

Merge s into t.

fn virtual_merge( &self, s: Group, t: Group, ea: &Self::Args, sc: &mut Self::Scratch, ) -> f64

Price a merge.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".